requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
Brotli==1.0.9
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
dataclasses-json==0.5.7
faiss-cpu==1.7.3
frozenlist==1.3.3
gptcache==0.1.11
greenlet==2.0.2
idna==3.4
langchain==0.0.141
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.24.2
openai==0.27.4
openapi-schema-pydantic==1.2.4
packaging==23.1
pydantic==1.10.7
PyPDF2==3.0.1
python-dotenv==1.0.0
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
soupsieve==2.4
SQLAlchemy==1.4.47
tenacity==8.2.2
tiktoken==0.3.3
tqdm==4.65.0
typing-inspect==0.8.0
typing_extensions==4.5.0
urllib3==1.26.15
yarl==1.8.2
